Add authorization
The authorization system is based on JWT token + email and password. A
user creates an account with an email and a password. The password is
safely stores encrypted using bcrypt. With valid credentials, users
receive a JWT, managed by Guardian. For requests that require
authorization, a header "Authorziation: Beared <JWT>" is required, and
Guardian will know whether to allow the user to get in.

Currently, the routes that require authorization are the ones to update
a producer, and to delete a producer.